[sim] more fp<->int fixes

+++ riscv/insns/fcvt_d_wu.h
@@ -1,4 +1,4 @@
 require_fp;
 softfloat_roundingMode = RM;
-FRD = ui32_to_f64(RS1);
+FRD = ui32_to_f64((uint32_t)RS1);
 set_fp_exceptions;
